Pizza RPG Battle System Part 1
Let's create a Battle System for our Pizza Legends game! In this video, we'll cover the overall setup and design of a turn based battle system. Our pizzas combatants will be set up to have health, stats, teams, and more. We'll also create a few HUD components that can be used throughout the project.